#1175 Correction to r4756 Another three casts required to build with Visual Studio (size_t code ripple)
(cherry picked from commit 44da4724b38935c672dd683484b4a575b600d1d4)

+++ src/preview.cpp
@@ -470,7 +470,7 @@ namespace {
         }
         IoCloser closer(io);
         const byte* data = io.mmap();
-        if (io.size() < nativePreview_.position_ + static_cast<long>(nativePreview_.size_)) {
+        if ((long)io.size() < nativePreview_.position_ + static_cast<long>(nativePreview_.size_)) {
 #ifndef SUPPRESS_WARNINGS
             EXV_WARNING << "Invalid native preview position or size.\n";
 #endif
